fix: Camera does not start after permission is granted
Using state to include the camera component after check permission is performed, this fixes the problem of camera not starting when the user first open the app and grant permission.

Another benefit is now the camera is actually turned of when face detection is disabled, by removing the camera component.